working on interface to GMP
We no longer set the GMP memory functions with mp_set_memory_functions.
This ensures that libraries that use GMP don't mislay our GC pointers.

To fix it, we have to finalize every top level object that might contain
a pointer to a GMP integer.  We hope this won't be too many
finalizations -- this is an experiment to see.

Currently I'm working on the interpreter, not the engine, being the only
place where top level GMP integers themselves are allocated and finalized.